- Removed Home.jsx - Added new folders - Organised imports
