+++ TaskManagementSystem/client/src/App.jsx
@@ -1,21 +1,19 @@
-import React, { Component } from 'react';
-import { BrowserRouter as Router, Route, withRouter, Switch } from 'react-router-dom';
 import { LoginCallback } from '@okta/okta-react';
-import Home from './pages/Home';
-import Login from './pages/Login';
-import Profile from './pages/Profile';
-import ProjectList from './pages/ProjectList';
-import NewProjectForm from './pages/NewProjectForm';
+import React, { Component } from 'react';
+import { Route, BrowserRouter as Router, Switch, withRouter } from 'react-router-dom';
 import TitleBar from './components/TitleBar';
-import PrivateRoute from './routes/PrivateRoute';
 import { AuthProvider } from './contexts/AuthContext';
+import NewProjectForm from './pages/HomePage/NewProjectForm';
+import ProjectList from './pages/HomePage/HomeView';
+import Login from './pages/Login/Login';
+import Profile from './pages/Debugging/Profile';
+import PrivateRoute from './routes/PrivateRoute';
 
 function App() {
   return (
     <Router>
       <TitleBar toggleNavbar={() => {}} />
       <Switch>
-        <Route path="/general" component={Home}/>
         <AuthProvider>
           <Route path="/" exact={true} component={Login}/>
           <Route path="/login/callback" component={LoginCallback}/>
