+++ TaskManagementSystem/client/src/pages/Debugging/Profile.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import { useHistory } from 'react-router-dom'
-import { AuthContext } from '../contexts/AuthContext';
+import { AuthContext } from '../../contexts/AuthContext';
 
 const Profile = () => {
   const { user, oktaAuth, auth } = useContext(AuthContext);
